refactor(commonTests): split up helpers into individual modules
