+++ packages/calcite-components/src/tests/commonTests.ts
@@ -1,683 +1,28 @@
 /* eslint-disable jest/no-conditional-expect -- Using conditional logic in a confined test helper to handle specific scenarios, reducing duplication, balancing test readability and maintainability. **/
 /* eslint-disable jest/no-export -- Util functions are now imported to be used as `it` blocks within `describe` instead of assertions within `it` blocks. */
 import { E2EElement, E2EPage, EventSpy, newE2EPage } from "@stencil/core/testing";
-import axe from "axe-core";
 import { toHaveNoViolations } from "jest-axe";
-import { config } from "../../stencil.config";
+import { KeyInput } from "puppeteer";
 import { html } from "../../support/formatting";
 import type { JSX } from "../components";
 import { getClearValidationEventName, hiddenFormInputSlotName, componentsWithInputEvent } from "../utils/form";
 import { MessageBundle } from "../utils/t9n";
+import { GlobalTestProps, IntrinsicElementsWithProp, newProgrammaticE2EPage, skipAnimations } from "./utils";
 import {
-  GlobalTestProps,
-  IntrinsicElementsWithProp,
-  isElementFocused,
-  newProgrammaticE2EPage,
-  skipAnimations,
-} from "./utils";
-import { KeyInput } from "puppeteer";
+  TagOrHTMLWithBeforeContent,
+  isHTML,
+  getTag,
+  simplePageSetup,
+  getTagAndPage,
+  ComponentTestSetup,
+  getTagOrHTMLWithBeforeContent,
+} from "./commonTestUtils/setupForTests";
 
 expect.extend(toHaveNoViolations);
 
 type ComponentTag = keyof JSX.IntrinsicElements;
-type AxeOwningWindow = GlobalTestProps<{ axe: typeof axe }>;
 type ComponentHTML = string;
 type TagOrHTML = ComponentTag | ComponentHTML;
-type BeforeContent = (page: E2EPage) => Promise<void>;
-
-export type TagAndPage = {
-  tag: ComponentTag;
-  page: E2EPage;
-};
-
-type TagOrHTMLWithBeforeContent = {
-  tagOrHTML: TagOrHTML;
-
-  /**
-   * Allows for custom setup of the page.
-   *
-   * This is useful for test helpers that need to create and configure the test page before running tests.
-   *
-   * @param page
-   */
-  beforeContent: BeforeContent;
-};
-
-export const HYDRATED_ATTR = config.hydratedFlag.name;
-
-function isHTML(tagOrHTML: string): boolean {
-  return tagOrHTML.trim().startsWith("<");
-}
-
-function getTag(tagOrHTML: string): ComponentTag {
-  if (isHTML(tagOrHTML)) {
-    const calciteTagRegex = /<calcite-[a-z0-9-]+/i;
-    const trimmedTag = tagOrHTML.trim();
-    const calciteTagMatchResult = trimmedTag.match(calciteTagRegex);
-    if (calciteTagMatchResult) {
-      return calciteTagMatchResult[0].substring(1) as ComponentTag;
-    }
-    throw new Error(`Could not extract tag from HTML: ${trimmedTag}`);
-  }
-
-  return tagOrHTML as ComponentTag;
-}
-
-async function simplePageSetup(componentTagOrHTML: TagOrHTML): Promise<E2EPage> {
-  const componentTag = getTag(componentTagOrHTML);
-  const page = await newE2EPage({
-    html: isHTML(componentTagOrHTML) ? componentTagOrHTML : `<${componentTag}></${componentTag}>`,
-    failOnConsoleError: true,
-  });
-  await page.waitForChanges();
-
-  return page;
-}
-
-/**
- * Helper for asserting that a component is accessible.
- *
- * Note that this helper should be used within a describe block.
- *
- * @example
- * describe("accessible"), () => {
- *    accessible(`<calcite-tree></calcite-tree>`);
- * });
- *
- * @param {ComponentTestSetup} componentTestSetup - A component tag, html, or the tag and e2e page for setting up a test
- */
-export function accessible(componentTestSetup: ComponentTestSetup): void {
-  it("is accessible", async () => {
-    const { page, tag } = await getTagAndPage(componentTestSetup);
-
-    await page.addScriptTag({ path: require.resolve("axe-core") });
-    await page.waitForFunction(() => (window as AxeOwningWindow).axe);
-
-    expect(
-      await page.evaluate(async (componentTag: ComponentTag) => (window as AxeOwningWindow).axe.run(componentTag), tag),
-    ).toHaveNoViolations();
-  });
-}
-
-/**
- * Note that this helper should be used within a describe block.
- *
- * @example
- * describe("renders", () => {
- *    renders(`<calcite-tree></calcite-tree>`);
- * });
- * @param componentTestSetup
- *
- * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
- * @param {object} options - additional options to assert
- * @param {string} options.visible - is the component visible
- * @param {string} options.display - is the component's display "inline"
- */
-export async function renders(
-  componentTestSetup: ComponentTestSetup,
-  options?: {
-    visible?: boolean;
-    display: string;
-  },
-): Promise<void> {
-  it(`renders`, async () => {
-    const { page, tag } = await getTagAndPage(componentTestSetup);
-    const element = await page.find(tag);
-
-    expect(element).toHaveAttribute(HYDRATED_ATTR);
-    expect(await element.isVisible()).toBe(options?.visible ?? true);
-    expect((await element.getComputedStyle()).display).toBe(options?.display ?? "inline");
-  });
-}
-
-/**
- *
- * Helper for asserting that a component reflects
- *
- * Note that this helper should be used within a describe block.
- *
- * @example
- * describe("reflects", () => {
- *    reflects("calcite-action-bar", [
- *      {
- *        propertyName: "expandDisabled",
- *        value: true
- *      },
- *      {
- *        propertyName: "expanded",
- *        value: true
- *      }
- *    ])
- * })
- *
- * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
- * @param componentTestSetup
- * @param {object[]} propsToTest - the properties to test
- * @param {string} propsToTest.propertyName - the property name
- * @param {any} propsToTest.value - the property value (if boolean, needs to be `true` to ensure reflection)
- */
-export function reflects(
-  componentTestSetup: ComponentTestSetup,
-  propsToTest: {
-    propertyName: string;
-    value: any;
-  }[],
-): void {
-  const cases = propsToTest.map(({ propertyName, value }) => [propertyName, value]);
-
-  it.each(cases)("%p", async (propertyName, value) => {
-    const { page, tag: componentTag } = await getTagAndPage(componentTestSetup);
-    await skipAnimations(page);
-    const element = await page.find(componentTag);
-
-    const attrName = propToAttr(propertyName);
-    const componentAttributeSelector = `${componentTag}[${attrName}]`;
-
-    element.setProperty(propertyName, value);
-    await page.waitForChanges();
-
-    expect(await page.find(componentAttributeSelector)).toBeTruthy();
-
-    if (typeof value === "boolean") {
-      const getExpectedValue = (propValue: boolean): string | null => (propValue ? "" : null);
-      const negated = !value;
-
-      element.setProperty(propertyName, negated);
-      await page.waitForChanges();
-
-      expect(element.getAttribute(attrName)).toBe(getExpectedValue(negated));
-
-      element.setProperty(propertyName, value);
-      await page.waitForChanges();
-
-      expect(element.getAttribute(attrName)).toBe(getExpectedValue(value));
-    }
-  });
-}
-
-function propToAttr(name: string): string {
-  return name.replace(/([a-z0-9])([A-Z])/g, "$1-$2").toLowerCase();
-}
-
-/**
- * Helper for asserting that a property's value is its default
- *
- * Note that this helper should be used within a describe block.
- *
- * @example
- * describe("defaults", () => {
- *    defaults("calcite-action", [
- *      {
- *        propertyName: "active",
- *        defaultValue: false
- *      },
- *      {
- *        propertyName: "appearance",
- *        defaultValue: "solid"
- *      }
- *    ])
- * })
- *
- * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
- * @param componentTestSetup
- * @param {object[]} propsToTest - the properties to test
- * @param {string} propsToTest.propertyName - the property name
- * @param {any} propsToTest.value - the property value
- */
-export function defaults(
-  componentTestSetup: ComponentTestSetup,
-  propsToTest: {
-    propertyName: string;
-    defaultValue: any;
-  }[],
-): void {
-  it.each(propsToTest.map(({ propertyName, defaultValue }) => [propertyName, defaultValue]))(
-    "%p",
-    async (propertyName, defaultValue) => {
-      const { page, tag } = await getTagAndPage(componentTestSetup);
-      const element = await page.find(tag);
-      const prop = await element.getProperty(propertyName);
-      expect(prop).toEqual(defaultValue);
-    },
-  );
-}
-
-/**
- * Helper for asserting that a component is not visible when hidden
- *
- * Note that this helper should be used within a describe block.
- *
- * @example
- * @param componentTestSetup
- * describe("honors hidden attribute", () => {
- *    hidden("calcite-accordion")
- * });
- *
- * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
- */
-export async function hidden(componentTestSetup: ComponentTestSetup): Promise<void> {
-  it("is hidden", async () => {
-    const { page, tag } = await getTagAndPage(componentTestSetup);
-    const element = await page.find(tag);
-
-    element.setAttribute("hidden", "");
-    await page.waitForChanges();
-
-    expect(await element.isVisible()).toBe(false);
-  });
-}
-
-interface FocusableOptions {
-  /**
-   * use this to pass an ID to setFocus()
-   *
-   * @deprecated components should no longer use a focusId parameter for setFocus()
-   */
-  focusId?: string;
-
-  /**
-   * selector used to assert the focused DOM element
-   */
-  focusTargetSelector?: string;
-
-  /**
-   * selector used to assert the focused shadow DOM element
-   */
-  shadowFocusTargetSelector?: string;
-}
-
-/**
- * Helper for asserting that a component is focusable
- *
- * Note that this helper should be used within a describe block.
- *
- * @example
- * describe("is focusable", () => {
- *    focusable(`calcite-input-number`, { shadowFocusTargetSelector: "input" })
- * });
- *
- * @param {string} componentTagOrHTML - the component tag or HTML markup to test against
- * @param componentTestSetup
- * @param {FocusableOptions} [options] - additional options for asserting focus
- */
-export function focusable(componentTestSetup: ComponentTestSetup, options?: FocusableOptions): void {
-  it("is focusable", async () => {
-    const { page, tag } = await getTagAndPage(componentTestSetup);
-    const element = await page.find(tag);
-    const focusTargetSelector = options?.focusTargetSelector || tag;
-    await element.callMethod("setFocus", options?.focusId); // assumes element is FocusableElement
-
-    if (options?.shadowFocusTargetSelector) {
-      expect(
-        await page.$eval(
-          tag,
-          (element: HTMLElement, selector: string) => element.shadowRoot.activeElement?.matches(selector),
-          options?.shadowFocusTargetSelector,
-        ),
-      ).toBe(true);
-    }
-
-    // wait for next frame before checking focus
-    await page.waitForChanges();
-
-    expect(await page.evaluate((selector) => document.activeElement?.matches(selector), focusTargetSelector)).toBe(
-      true,
-    );
-  });
-}
-
-/**
- * Helper for asserting slots.
- *
- * Note that this helper should be used within a describe block.
- *
- * @example
- * describe("slots", () => {
- *    slots("calcite-stack", SLOTS)
- * })
- *
- * @param {string} componentTagOrHTML - The component tag or HTML markup to test against.
- * @param {Record<string, string> | string[]} slots - A component's SLOTS resource object or an array of slot names.
- * @param {boolean} includeDefaultSlot - When true, it will run assertions on the default slot.
- */
-export function slots(
-  componentTagOrHTML: TagOrHTML,
-  slots: Record<string, string> | string[],
-  includeDefaultSlot = false,
-): void {
-  it("has slots", async () => {
-    const page = await simplePageSetup(componentTagOrHTML);
-    const tag = getTag(componentTagOrHTML);
-    const slotNames = Array.isArray(slots) ? slots : Object.values(slots);
-
-    await page.$eval(
-      tag,
-      async (component, slotNames: string[], includeDefaultSlot?: boolean) => {
-        async function slotTestElement(testClass: string, slotName?: string): Promise<void> {
-          const el = document.createElement("div"); // slotting a <div> will suffice for our purposes
-          el.classList.add(testClass);
-
-          if (slotName) {
-            el.slot = slotName;
-          }
-
-          component.append(el);
-          await new Promise<void>((resolve) => requestAnimationFrame(() => resolve()));
-        }
-
-        for (let i = 0; i < slotNames.length; i++) {
-          await slotTestElement("slotted-into-named-slot", slotNames[i]);
-        }
-
-        if (includeDefaultSlot) {
-          await slotTestElement("slotted-into-default-slot");
-        }
-      },
-      slotNames,
-      includeDefaultSlot,
-    );
-
-    await page.waitForChanges();
-
-    const slotted = await page.evaluate(() =>
-      Array.from(document.querySelectorAll(".slotted-into-named-slot"))
-        .filter((slotted) => slotted.assignedSlot)
-        .map((slotted) => slotted.slot),
-    );
-
-    expect(slotNames).toEqual(slotted);
-
-    if (includeDefaultSlot) {
-      const hasDefaultSlotted = await page.evaluate(() => {
-        const defaultSlotted = document.querySelector(".slotted-into-default-slot");
-        return defaultSlotted.assignedSlot?.name === "" && defaultSlotted.slot === "";
-      });
-
-      expect(hasDefaultSlotted).toBe(true);
-    }
-  });
-}
-
-async function assertLabelable({
-  page,
-  componentTag,
-  propertyToToggle,
-  focusTargetSelector = componentTag,
-  shadowFocusTargetSelector,
-}: {
-  page: E2EPage;
-  componentTag: string;
-  propertyToToggle?: string;
-  focusTargetSelector?: string;
-  shadowFocusTargetSelector?: string;
-}): Promise<void> {
-  let initialPropertyValue: boolean;
-  const component = await page.find(componentTag);
-
-  if (propertyToToggle) {
-    initialPropertyValue = await component.getProperty(propertyToToggle);
-  }
-
-  const label = await page.find("calcite-label");
-  await label.callMethod("click"); // we call the method to avoid clicking the child element
-  await page.waitForChanges();
-
-  expect(
-    await page.evaluate(
-      (focusTargetSelector: string): boolean => !!document.activeElement?.closest(focusTargetSelector),
-      focusTargetSelector,
-    ),
-  ).toBe(true);
-
-  if (shadowFocusTargetSelector) {
-    expect(
-      await page.$eval(
-        componentTag,
-        (element: HTMLElement, selector: string) => element.shadowRoot.activeElement.matches(selector),
-        shadowFocusTargetSelector,
-      ),
-    ).toBe(true);
-  }
-
-  if (propertyToToggle) {
-    const toggledPropertyValue = !initialPropertyValue;
-    expect(await component.getProperty(propertyToToggle)).toBe(toggledPropertyValue);
-
-    // assert that direct clicks on component toggle property correctly
-    component.setProperty(propertyToToggle, initialPropertyValue); // we reset as not all components toggle when clicked
-    await page.waitForChanges();
-    await component.click();
-    await page.waitForChanges();
-    expect(await component.getProperty(propertyToToggle)).toBe(toggledPropertyValue);
-  }
-
-  // assert clicking on labelable keeps focus
-  await component.callMethod("click");
-  await page.waitForChanges();
-
-  expect(await isElementFocused(page, focusTargetSelector)).toBe(true);
-}
-
-interface LabelableOptions extends Pick<FocusableOptions, "focusTargetSelector" | "shadowFocusTargetSelector"> {
-  /**
-   * If clicking on a label toggles the labelable component, use this prop to specify the name of the toggled prop.
-   */
-  propertyToToggle?: string;
-}
-
-/**
- * Helper for asserting label clicking functionality works.
- *
- * Note that this helper should be used within a describe block.
- *
- * @example
- * describe("labelable", () => {
- *    async () => labelable("calcite-button")
- * })
- *
- * @param {string} componentTagOrHtml - The component tag or HTML used to test label support.
- * @param {LabelableOptions} [options] - Labelable options.
- */
-export function labelable(
-  componentTagOrHtml: TagOrHTML | TagOrHTMLWithBeforeContent,
-  options?: LabelableOptions,
-): void {
-  const id = "labelable-id";
-  const labelTitle = "My Component";
-  const propertyToToggle = options?.propertyToToggle;
-  const focusTargetSelector = options?.focusTargetSelector || `#${id}`;
-  const shadowFocusTargetSelector = options?.shadowFocusTargetSelector;
-  const { beforeContent, tagOrHTML } = getTagOrHTMLWithBeforeContent(componentTagOrHtml);
-  const componentTag = getTag(tagOrHTML);
-  const componentHtml = isHTML(tagOrHTML) ? ensureId(tagOrHTML) : `<${componentTag} id="${id}"></${componentTag}>`;
-
-  function ensureId(html: string): string {
-    return html.includes("id=") ? html : html.replace(componentTag, `${componentTag} id="${id}" `);
-  }
-
-  describe("label wraps labelables", () => {
-    it("is labelable when component is wrapped in a label", async () => {
-      const wrappedHtml = html`<calcite-label>${labelTitle} ${componentHtml}</calcite-label>`;
-      const wrappedPage: E2EPage = await newE2EPage();
-      beforeContent?.(wrappedPage);
-      await wrappedPage.setContent(wrappedHtml);
-      await wrappedPage.waitForChanges();
-
-      await assertLabelable({
-        page: wrappedPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector,
-        shadowFocusTargetSelector,
-      });
-    });
-
-    it("is labelable when wrapping label is set prior to component", async () => {
-      const labelFirstWrappedPage: E2EPage = await newE2EPage();
-      beforeContent?.(labelFirstWrappedPage);
-      await labelFirstWrappedPage.setContent(html`
-        <calcite-label></calcite-label>
-        <template>${componentHtml}</template>
-      `);
-      await labelFirstWrappedPage.waitForChanges();
-      await labelFirstWrappedPage.evaluate(() => {
-        const template = document.querySelector("template");
-        const labelEl = document.querySelector("calcite-label");
-
-        labelEl.append(template.content.cloneNode(true));
-      }, componentHtml);
-      await labelFirstWrappedPage.waitForChanges();
-
-      await assertLabelable({
-        page: labelFirstWrappedPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector,
-        shadowFocusTargetSelector,
-      });
-    });
-
-    it("is labelable when a component is set first before being wrapped in a label", async () => {
-      const componentFirstWrappedPage: E2EPage = await newE2EPage();
-      beforeContent?.(componentFirstWrappedPage);
-      await componentFirstWrappedPage.setContent(componentHtml);
-      await componentFirstWrappedPage.waitForChanges();
-      await componentFirstWrappedPage.evaluate((id: string) => {
-        const componentEl = document.querySelector(`[id='${id}']`);
-        const labelEl = document.createElement("calcite-label");
-        document.body.append(labelEl);
-        labelEl.append(componentEl);
-      }, id);
-      await componentFirstWrappedPage.waitForChanges();
-
-      await assertLabelable({
-        page: componentFirstWrappedPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector,
-        shadowFocusTargetSelector,
-      });
-    });
-
-    it("only sets focus on the first labelable when label is clicked", async () => {
-      const firstLabelableId = `${id}`;
-      const componentFirstWrappedPage: E2EPage = await newE2EPage();
-      beforeContent?.(componentFirstWrappedPage);
-      const content = html`
-        <calcite-label>
-          <!-- duplicate tags should be fine as assertion uses first match -->
-          ${componentHtml.replace(id, firstLabelableId)} ${componentHtml.replace(id, `${id}-2`)}
-          ${componentHtml.replace(id, `${id}-3`)}
-        </calcite-label>
-      `;
-
-      await componentFirstWrappedPage.setContent(content);
-      await componentFirstWrappedPage.waitForChanges();
-
-      const firstLabelableSelector = `#${firstLabelableId}`;
-
-      await assertLabelable({
-        page: componentFirstWrappedPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector:
-          focusTargetSelector === firstLabelableSelector
-            ? firstLabelableSelector
-            : `${firstLabelableSelector} ${focusTargetSelector}`,
-      });
-    });
-  });
-
-  describe("label is sibling to labelables", () => {
-    it("is labelable with label set as a sibling to the component", async () => {
-      const siblingHtml = html`
-        <calcite-label for="${id}">${labelTitle}</calcite-label>
-        ${componentHtml}
-      `;
-      const siblingPage: E2EPage = await newE2EPage();
-      beforeContent?.(siblingPage);
-
-      await siblingPage.setContent(siblingHtml);
-      await siblingPage.waitForChanges();
-
-      await assertLabelable({
-        page: siblingPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector,
-        shadowFocusTargetSelector,
-      });
-    });
-
-    it("is labelable when sibling label is set prior to component", async () => {
-      const labelFirstSiblingPage: E2EPage = await newE2EPage();
-      beforeContent?.(labelFirstSiblingPage);
-      await labelFirstSiblingPage.setContent(html`
-        <calcite-label for="${id}"></calcite-label>
-        <template>${componentHtml}</template>
-      `);
-      await labelFirstSiblingPage.waitForChanges();
-      await labelFirstSiblingPage.evaluate(() => {
-        const template = document.querySelector("template");
-        document.body.append(template.content.cloneNode(true));
-      }, componentHtml);
-      await labelFirstSiblingPage.waitForChanges();
-
-      await assertLabelable({
-        page: labelFirstSiblingPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector,
-        shadowFocusTargetSelector,
-      });
-    });
-
-    it("is labelable for a component set before sibling label", async () => {
-      const componentFirstSiblingPage: E2EPage = await newE2EPage();
-      beforeContent?.(componentFirstSiblingPage);
-      await componentFirstSiblingPage.setContent(componentHtml);
-      await componentFirstSiblingPage.waitForChanges();
-      await componentFirstSiblingPage.evaluate((id: string) => {
-        const label = document.createElement("calcite-label");
-        label.setAttribute("for", `${id}`);
-        document.body.append(label);
-      }, id);
-      await componentFirstSiblingPage.waitForChanges();
-
-      await assertLabelable({
-        page: componentFirstSiblingPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector,
-        shadowFocusTargetSelector,
-      });
-    });
-
-    it("is labelable when label's for is set after initialization", async () => {
-      const siblingHtml = html`
-        <calcite-label>${labelTitle}</calcite-label>
-        ${componentHtml}
-      `;
-      const siblingPage: E2EPage = await newE2EPage();
-      beforeContent?.(siblingPage);
-
-      await siblingPage.setContent(siblingHtml);
-      await siblingPage.waitForChanges();
-
-      const label = await siblingPage.find("calcite-label");
-      label.setProperty("for", id);
-      await siblingPage.waitForChanges();
-
-      await assertLabelable({
-        page: siblingPage,
-        componentTag,
-        propertyToToggle,
-        focusTargetSelector,
-        shadowFocusTargetSelector,
-      });
-    });
-  });
-}
 
 interface FormAssociatedOptions {
   /**
@@ -1141,39 +486,6 @@ interface DisabledOptions {
    *  Note: this should only be used for components that wrap a single component that implements disabled behavior.
    */
   shadowAriaAttributeTargetSelector?: string;
-}
-
-type ComponentTestContent = TagOrHTML | TagAndPage;
-type ComponentTestSetupProvider = (() => ComponentTestContent) | (() => Promise<ComponentTestContent>);
-type ComponentTestSetup = ComponentTestContent | ComponentTestSetupProvider;
-
-async function getTagAndPage(componentTestSetup: ComponentTestSetup): Promise<TagAndPage> {
-  if (typeof componentTestSetup === "function") {
-    componentTestSetup = await componentTestSetup();
-  }
-
-  if (typeof componentTestSetup === "string") {
-    const page = await simplePageSetup(componentTestSetup);
-    const tag = getTag(componentTestSetup);
-
-    return { page, tag };
-  }
-
-  return componentTestSetup;
-}
-
-function getTagOrHTMLWithBeforeContent(componentTestSetup: TagOrHTML | TagOrHTMLWithBeforeContent): {
-  tagOrHTML: TagOrHTML;
-  beforeContent?: BeforeContent;
-} {
-  if (typeof componentTestSetup === "string") {
-    return { tagOrHTML: componentTestSetup };
-  }
-
-  return {
-    tagOrHTML: componentTestSetup.tagOrHTML,
-    beforeContent: componentTestSetup.beforeContent,
-  };
 }
 
 /**
